requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
python-multipart
pypdf
faiss-cpu
sentence-transformers

